fix CF
